retrieve cumulative activities from backend
# Conflicts:
#	src/state/applet/applet.actions.js
#	src/widgets/ActivitySummary.js
