+++ src/state/applet/applet.actions.js
@@ -6,6 +6,7 @@ import { responsesSelector } from '../app/app.selectors';
 import { appletsSelector } from './applet.selectors';
 import { updateKeys } from '../responses/responses.actions';
 import { replaceResponses } from '../responses/responses.reducer';
+import { setCumulativeActivities } from './applet.reducer';
 
 import { transformApplet, parseAppletEvents } from '../../services/json-ld';
 import { getAppletsAPI, getPublicAppletAPI } from '../../services/applet.service';
@@ -27,8 +28,23 @@ export const getApplets = createAsyncThunk(APPLET_CONSTANTS.GET_APPLETS, async (
   const applets = await getAppletsAPI({ token, localInfo });
 
   const transformedApplets = [];
+  let cumulativeActivities = {};
+
+  const getCumulativeActivities = (applet, nextActivities) => {
+    const response = {};
+    for (const activityId in nextActivities) {
+      response[`activity/${activityId}/nextActivity`] = nextActivities[activityId].map(id => {
+        const activity = applet.activities.find(activity => activity.id.split('/').pop() == id)
+        return activity && activity.name.en;
+      }).filter(name => name.length)
+    }
+
+    return response;
+  }
+
   for (let index = 0; index < applets.data.length; index++) {
     const appletInfo = applets.data[index];
+    const nextActivities = appletInfo.cumulativeActivities;
 
     if (!appletInfo.applet) {
       const applet = modifyApplet(appletInfo, currentApplets);
@@ -53,6 +69,7 @@ export const getApplets = createAsyncThunk(APPLET_CONSTANTS.GET_APPLETS, async (
         });
 
         transformedApplets.push(applet)
+        Object.assign(cumulativeActivities, getCumulativeActivities(applet, nextActivities))
       }
     } else {
       const applet = transformApplet(appletInfo, currentApplets);
@@ -80,10 +97,12 @@ export const getApplets = createAsyncThunk(APPLET_CONSTANTS.GET_APPLETS, async (
           appletId: 'applet/' + appletInfo.id
         });
         transformedApplets.push(applet);
+        Object.assign(cumulativeActivities, getCumulativeActivities(applet, nextActivities))
       }
     }
   };
 
+  dispatch(setCumulativeActivities(cumulativeActivities));
   dispatch(replaceResponses(responses));
 
   return transformedApplets;
